Additional Changes for Standalone ISO using Initializer
Initializer is used in lieu of Installer as most installer
functions are not meaningful when installing from
a local repository (If the installer every does something
more sophisticated than yum install 128T, the Installer,
rather than kickstart will have to be used in the standalone
ISO to do the install.

Two options are provided:
- Install by Serial Console, /dev/ttySN
- Install by Standard (i.e. Graphical Console).. /dev/tty1

It may be possible to merge the serial and graphical installs
by using adding console=/dev/ttySN... to a single grub menu entry
and altering the scripts used to invoke the initializer, but
VirtualBox (the primary unit test environment) does not offer
a serial console and will not run anaconda when console= is added
to the bootline (instead the screen is filled with dots and the
install hangs).

YMM with other hypervisors...

Serial and graphical getty manipulation via systemd services
used to run initializer before first console login.

A snippet is provided to construct the local repo required
by install128t run in the post-install environment, but so
long as the Initializer alone is used, there is no need to
waste time copying files from the install media local
repo to the post-install environment.

A few new packages were added to the 128t.cfg manifest.