Change pkglast parameter to list of packages
In order to invoke mkiso.sh from a make target, the 128T rpm needs to
be specified fom the commandlind without replacing the other rpms
already in the list.

This change makes pkglast as a second list of packages to install,
rather than a regex to be run against the only package list to
produce the second list as was its former behavior.