Add mkiso.sh --rpm_to_iso_pattern parameter
(1) Add --rpm-to-iso-pattern parameter.  The parameter value is
a regex (e.g. 128T-[1-9]) which matches packages specified in
the --pkglist or --pkglast parameter value lists. The filename
of the matching rpm has the .rpm extension stripped and replaced
with a .iso extension.  The match is performed after any
transformations specified in --pkglist or --pkglast (so that
~128T might be relaced with 128T-3.1.10-1.el7.centos.x86_64.rpm
and be changed to 128T-3.1.10-1.el7.centos.x86_64.iso). The output
directory for the ISO is determined from the --iso-out parameter
value.  If that parameter value specifies the filename, the
path is still selected but the filename is replaced.

(2) Address problem with glob expansion of items passed to
add_to_list() function.

(3) Address problem with processing of absolute path parmeter
values

(4) Cleanup --pkglast procesing in yum_download() function.