Add mkiso support for input rpm file list
Turns out this was already done for reconstructing an installable
ISO from a save tech-support-info RPM list.

The 'mkiso rpm2iso' accepts an argument --rpm-file=<filename> which
expects a list of rpm versions or rpm filenames. This is compatible
with the format of the pincusion file.

The only change required was updating the rpm2iso command to process
--iso-in=http://...

BEL-33