Add 128T-manifest support to mkiso
(1) Transform the specified 128T package name into a corresponding
128T-manifest name, installs 128T-manifest into the yum virtual root
extacts the packge list file and adds to the package list.

(2) Add support for alternate paths to look for copy source files
in.  This will allow for common script and snippet paths so these files
do not have to be unnecessarily duplicated for new ISO config profiles.